Move New Syndication APIs to NetCore Specific Ref
Move New Syndication APIs to NetCore Specific Ref Assembly.
Fix #25585
