+++ src/System.ServiceModel.Syndication/tests/netcoreapp/Configurations.props
@@ -2,9 +2,7 @@
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <BuildConfigurations>
-      netfx;
       netcoreapp;
-      netstandard;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
